requirements.txt
autopep8>=1.4.4
backcall>=0.1.0
beautifulsoup4>=4.8.1
bs4>=0.0.1
certifi>=2019.11.28
cffi>=1.13.2
chardet>=3.0.4
colorama>=0.4.1
cryptography>=2.8
dateutils>=0.6.7
decorator>=4.4.1
dotmap>=1.3.8
html5lib>=1.0.1
idna>=2.8
ipdb>=0.12.2
ipython>=7.9.0
ipython-genutils>=0.2.0
jedi>=0.15.1
lxml>=4.4.2
parso>=0.5.1
pexpect>=4.7.0
pickleshare>=0.7.5
play-scraper>=0.6.0
PrettyTable>=0.7.2
prompt-toolkit>=2.0.10
ptyprocess>=0.6.0
py-backwards>=0.7
py-backwards-astunparse>=1.5.0.post3
pycodestyle>=2.5.0
pycparser>=2.19
Pygments>=2.5.1
pyOpenSSL>=19.1.0
python-dateutil>=2.8.1
pytz>=2019.3
requests>=2.22.0
requests-futures>=1.0.0
six>=1.13.0
soupsieve>=1.9.5
traitlets>=4.3.3
typed-ast>=1.4.0
urllib3>=1.25.7
wcwidth>=0.1.7
webencodings>=0.5.1
